chore: use a central code location to check PHP version
