+++ console.php
@@ -33,28 +33,6 @@ use Symfony\Component\Console\Output\ConsoleOutput;
 
 \define('OC_CONSOLE', 1);
 
-// Show warning if a PHP version below 7.4.0 is used, this has to happen here
-// because base.php will already use 7.4 syntax.
-if (\version_compare(PHP_VERSION, '7.4.0') === -1) {
-	echo 'This version of ownCloud requires at least PHP 7.4.0'.PHP_EOL;
-	echo 'You are currently running PHP ' . PHP_VERSION . '. Please update your PHP version.'.PHP_EOL;
-	exit(1);
-}
-
-// Show warning if PHP 8.0 or later is used as ownCloud is not compatible with PHP 8.0
-if (\version_compare(PHP_VERSION, '8.0.0') !== -1) {
-	echo 'This version of ownCloud is not compatible with PHP 8.0' . PHP_EOL;
-	echo 'You are currently running PHP ' . PHP_VERSION . '.' . PHP_EOL;
-	exit(1);
-}
-
-// running oC on Windows is unsupported since 8.1, this has to happen here because
-// is seems that the autoloader on Windows fails later and just throws an exception.
-if (\stripos(PHP_OS, 'WIN') === 0) {
-	echo 'ownCloud Server does not support Microsoft Windows.';
-	exit(1);
-}
-
 function exceptionHandler($exception) {
 	try {
 		// try to log the exception
